refactor(inference): carve the SSE stream loop and its wire types out of generate_stream (S3b, pure code-motion)
`inference/sse_stream.rs`: the stream consumer (queue-vs-decode watchdogs, prefill
progress as liveness, token / reasoning / tool-call / usage / timings accumulation) and
the wire types it parses (`OpenAIStreamChunk` family, `OpenAIUsage`, `OpenAITimings`,
`StreamToolAccum`, `accumulate_stream_tool_call`, `warn_if_decode_collapsed`,
`STREAM_IDLE_TIMEOUT_SECS`). The loop's accumulators come back as one `StreamOutcome`
the caller destructures into the same locals the inline loop bound, so the
post-processing tail is untouched. Behaviour-identical; adapter 3664 → 3128 lines
(from 4191 this morning). `generate_stream` is now: preamble → body → admission →
LoRA page-in → finish body → endpoint → guard → send → stream → post-process.
