+++ core/continuum-core/src/inference/mod.rs
@@ -51,6 +51,7 @@ pub mod llama_server;
 pub mod llamacpp_adapter;
 pub mod measured_hold;
 pub mod slots;
+pub mod sse_stream;
 pub mod llm_module;
 pub mod llm_module_bus;
 pub mod llm_module_service;
